Organized a lot of the files, and added all Tier 2 Major Arcana to the translation file
