+++ src/main/java/net/luckofthedraw/item/ModItems.java
@@ -4,7 +4,7 @@ import net.fabricmc.fabric.api.item.v1.FabricItemSettings;
 import net.fabricmc.fabric.api.itemgroup.v1.FabricItemGroupEntries;
 import net.fabricmc.fabric.api.itemgroup.v1.ItemGroupEvents;
 import net.luckofthedraw.LuckOfTheDraw;
-import net.luckofthedraw.item.custom.*;
+import net.luckofthedraw.item.custom.majorArcana.*;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemGroups;
 import net.minecraft.registry.Registries;
