+++ src/main/resources/assets/luck_of_the_draw/lang/en_us.json
@@ -1,7 +1,8 @@
 {
-  "item.luck_of_the_draw.tarot_card.interact_fail": "This card doesn't seem ready yet.",
+  "luck_of_the_draw.misc.test": "§c<shake>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA<shake>",
+  "luck_of_the_draw.misc.wew": "<slow_wave>WEEEEEEEEEEEEEEEEEEEE<slow_wave>",
 
-  "item.luck_of_the_draw.misc.test": "§c<shake>AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA<shake>",
+  "item.luck_of_the_draw.tarot_card.interact_fail": "This card doesn't seem ready yet.",
 
   "item.luck_of_the_draw.the_fool_tarot_card": "The Fool",
   "item.luck_of_the_draw.the_fool_tarot_card.tooltip_type": "§7#0 of the §dMajor Arcana§7, §fT1",
@@ -24,15 +25,51 @@
   "item.luck_of_the_draw.the_emperor_tarot_card.tooltip_line1": "§7#4 of the §dMajor Arcana§7, §aT2",
   "item.luck_of_the_draw.the_emperor_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
 
+  "item.luck_of_the_draw.the_hierophant_tarot_card": "The Hierophant",
+  "item.luck_of_the_draw.the_hierophant_tarot_card.tooltip_line1": "§7#5 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.the_hierophant_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.the_lovers_tarot_card": "The Lovers",
+  "item.luck_of_the_draw.the_lovers_tarot_card.tooltip_line1": "§7#6 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.the_lovers_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.the_chariot_tarot_card": "The Chariot",
+  "item.luck_of_the_draw.the_chariot_tarot_card.tooltip_line1": "§7#7 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.the_chariot_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.strength_tarot_card": "Strength",
+  "item.luck_of_the_draw.strength_tarot_card.tooltip_line1": "§7#8 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.strength_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.the_hermit_tarot_card": "The Hermit",
+  "item.luck_of_the_draw.the_hermit_tarot_card.tooltip_line1": "§7#9 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.the_hermit_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.wheel_of_fortune_tarot_card": "Wheel of Fortune",
+  "item.luck_of_the_draw.wheel_of_fortune_tarot_card.tooltip_line1": "§7#10 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.wheel_of_fortune_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.justice_tarot_card": "Justice",
+  "item.luck_of_the_draw.justice_tarot_card.tooltip_line1": "§7#11 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.justice_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.the_hanged_man_tarot_card": "The Hanged Man",
+  "item.luck_of_the_draw.the_hanged_man_tarot_card.tooltip_line1": "§7#12 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.the_hanged_man_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
+  "item.luck_of_the_draw.death_tarot_card": "Death",
+  "item.luck_of_the_draw.death_tarot_card.tooltip_line1": "§7#13 of the §dMajor Arcana§7, §aT2",
+  "item.luck_of_the_draw.death_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
+
   "item.luck_of_the_draw.temperance_tarot_card": "Temperance",
   "item.luck_of_the_draw.temperance_tarot_card.tooltip_line1": "§7#14 of the §dMajor Arcana§7, §aT2§7-§bT3",
   "item.luck_of_the_draw.temperance_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
 
   "item.luck_of_the_draw.judgement_tarot_card": "Judgement",
-  "item.luck_of_the_draw.judgement_tarot_card.tooltip_line1": "§7#30 of the §dMajor Arcana§7, §bT3§7-§dT4",
+  "item.luck_of_the_draw.judgement_tarot_card.tooltip_line1": "§7#20 of the §dMajor Arcana§7, §bT3§7-§dT4",
   "item.luck_of_the_draw.judgement_tarot_card.tooltip_line2": "<shake>§cWIP<shake>",
 
   "item.luck_of_the_draw.the_world_tarot_card": "The World",
-  "item.luck_of_the_draw.the_world_tarot_card.tooltip_line1": "§7#31 of the §dMajor Arcana§7, §dT4",
+  "item.luck_of_the_draw.the_world_tarot_card.tooltip_line1": "§7#21 of the §dMajor Arcana§7, §dT4",
   "item.luck_of_the_draw.the_world_tarot_card.tooltip_line2": "<shake>§cWIP<shake>"
 }
